# feat: Optimize semantic chunking algorithm (O(n²) to O(n) complexity)