… src/shard_markdown/core/chunking/semantic_advanced.py
… tests/benchmarks/test_semantic_performance.py